# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Система управления учета оборудования и мебели",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/equipments": {
            "get": {
                "tags": [
                    "Equipments"
                ],
                "summary": "Получить список всего оборудования",
                "description": "Display a listing of the resource.",
                "operationId": "6ef4946ec56ce0f0c3e613f7c57d37f2",
                "responses": {
                    "200": {
                        "description": "Список оборудования",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/EquipmentIndexResource"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Equipments"
                ],
                "summary": "Создать новое оборудование",
                "description": "Store a newly created resource in storage.",
                "operationId": "c95b6d6ae06e995728be5a7e10fcbcc8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreEquipmentRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Оборудование создано",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EquipmentShowResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/equipments/{id}": {
            "get": {
                "tags": [
                    "Equipments"
                ],
                "summary": "Получить оборудование",
                "description": "Display the specified resource.",
                "operationId": "f5219eb4f4143cf304576feea563c147",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID оборудования",
                        "required": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Оборудование",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/EquipmentShowResource"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/equipments/{equipment}": {
            "delete": {
                "tags": [
                    "Equipments"
                ],
                "summary": "Удалить оборудование",
                "description": "Remove the specified resource from storage.",
                "operationId": "3503d7fc2ffb0a5ca3f23b76ff7c2f46",
                "parameters": [
                    {
                        "name": "equipment",
                        "in": "path",
                        "description": "ID оборудования",
                        "required": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Оборудование удалено"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Equipments"
                ],
                "summary": "Обновить оборудование",
                "description": "Update the specified resource in storage.",
                "operationId": "7111afb79e41bfb44428157d1d09c468",
                "parameters": [
                    {
                        "name": "equipment",
                        "in": "path",
                        "description": "ID оборудования",
                        "required": true,
                        "example": 1
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateEquipmentRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Оборудование изменен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EquipmentShowResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "StoreEquipmentRequest": {
                "title": "Запрос на создание оборудования",
                "description": "Данные для создания нового оборудования",
                "required": [
                    "equipment_model_id",
                    "cabinet_id",
                    "inventory_number"
                ],
                "properties": {
                    "equipment_model_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "cabinet_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "inventory_number": {
                        "type": "integer",
                        "example": 55555555
                    }
                },
                "type": "object"
            },
            "UpdateEquipmentRequest": {
                "title": "Запрос на обновление оборудования",
                "description": "Данные для обновления оборудования",
                "required": [
                    "equipment_model_id",
                    "cabinet_id",
                    "inventory_number"
                ],
                "properties": {
                    "equipment_model_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "cabinet_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "inventory_number": {
                        "type": "integer",
                        "example": 55555555
                    }
                },
                "type": "object"
            },
...
            "EquipmentIndexResource": {
                "title": "Ресурс списка оборудования",
                "description": "Формат данных оборудования при получении списка",
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "id": {
                                    "type": "integer",
                                    "example": 1
                                },
                                "equipment_model": {
                                    "type": "string",
                                    "example": "EXP-123"
                                },
                                "cabinet_id": {
                                    "type": "integer",
                                    "example": 1
                                },
                                "inventory_number": {
                                    "type": "integer",
                                    "example": 555555555
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
            "EquipmentShowResource": {
                "title": "Ресурс оборудования",
                "description": "Формат данных оборудования",
                "properties": {
                    "data": {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": 1
                            },
                            "equipment_model": {
                                "type": "string",
                                "example": "EXP-123"
                            },
                            "cabinet_number": {
                                "type": "integer",
                                "example": 321
                            },
                            "inventory_number": {
                                "type": "integer",
                                "example": 555555555
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}